[Major] Removed dlg and added reload()
This is so that the strolch-wc-auth component can be used in iron-pages where embedding a dialog leads to iron-backdrop being placed wrongly. The reload method is also used so that nothing is done before it needs to be, when embedded in iron-pages.

Now for strolch-wc-auth to work, the parent must set the app and app-title fields, incl. offer a dlg paper-dialog and two properties dlgTitle and dlgText. Then the parent must all the reload() method on the auth element to trigger the auth check.

To be notified of the auth check status the parent must then offer the method sessionValid(), so that the UI can be updated